+++ allennlp_plugins/declutr/__init__.py
@@ -1,0 +1,1 @@
+import declutr
