--- allennlp_plugins/t2t/__init__.py
@@ -1,1 +1,0 @@
-import t2t
