+++ declutr/__init__.py
@@ -1,0 +1,2 @@
+from declutr.declutr import DeCLUTR
+from declutr.encoder import Encoder
