+++ declutr/data/dataset_readers/__init__.py
@@ -1,0 +1,1 @@
+from declutr.data.dataset_readers.contrastive import ContrastiveDatasetReader
