+++ declutr/losses/__init__.py
@@ -1,0 +1,8 @@
+from declutr.losses.pytorch_metric_learning import (
+    CircleLoss,
+    ContrastiveLoss,
+    CrossBatchMemory,
+    MultiSimilarityLoss,
+    NTXentLoss,
+    PyTorchMetricLearningLoss,
+)
