+++ declutr/modules/text_field_embedders/__init__.py
@@ -1,0 +1,1 @@
+from declutr.modules.text_field_embedders.mlm_text_field_embedder import MLMTextFieldEmbedder
