+++ declutr/modules/token_embedders/__init__.py
@@ -1,0 +1,3 @@
+from declutr.modules.token_embedders.pretrained_transformer_embedder_mlm import (
+    PretrainedTransformerEmbedderMLM,
+)
