+++ declutr/predictors/__init__.py
@@ -1,0 +1,1 @@
+from declutr.predictors.contrastive_predictor import ContrastivePredictor
