+++ setup.py
@@ -4,22 +4,23 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="t2t",
+    name="declutr",
     version="0.1.0",
     author="John Giorgi",
     author_email="johnmgiorgi@gmail.com",
-    description=("A contrastive, self-supervised method for learning textual representations."),
+    description=("DeCLUTR: Deep Contrastive Learning for Unsupervised Textual Representations."),
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/JohnGiorgi/t2t",
+    url="https://github.com/JohnGiorgi/DeCLUTR",
     packages=setuptools.find_packages(),
     keywords=[
+        "universal sentence embeddings" "contrastive learning",
+        "representation learning",
+        "textual representations",
         "natural language processing",
-        "pytorch",
         "allennlp",
         "transformers",
-        "contrastive learning",
-        "textual representations",
+        "pytorch",
     ],
     classifiers=[
         "Development Status :: 1 - Planning",
@@ -34,6 +35,6 @@ setuptools.setup(
         "Typing :: Typed",
     ],
     python_requires=">=3.7",
-    install_requires=["torch>=1.4.0", "pytorch-metric-learning>=0.9.85", "typer>=0.2.1"],
+    install_requires=["torch>=1.5.0", "pytorch-metric-learning>=0.9.85", "typer>=0.2.1"],
     extras_require={"dev": ["black", "flake8", "hypothesis", "pytest"]},
 )
