--- t2t/__init__.py
@@ -1,1 +1,0 @@
-from t2t.encoder import Encoder
