--- t2t/data/dataset_readers/__init__.py
@@ -1,1 +1,0 @@
-from t2t.data.dataset_readers.contrastive import ContrastiveDatasetReader
