--- t2t/losses/__init__.py
@@ -1,8 +1,0 @@
-from t2t.losses.pytorch_metric_learning import (
-    CircleLoss,
-    ContrastiveLoss,
-    CrossBatchMemory,
-    MultiSimilarityLoss,
-    NTXentLoss,
-    PyTorchMetricLearningLoss,
-)
