--- t2t/miners/__init__.py
@@ -1,6 +1,0 @@
-from t2t.miners.pytorch_metric_learning import (
-    BatchHardMiner,
-    MultiSimilarityMiner,
-    PairMarginMiner,
-    PyTorchMetricLearningMiner,
-)
