--- t2t/modules/text_field_embedders/__init__.py
@@ -1,1 +1,0 @@
-from t2t.modules.text_field_embedders.mlm_text_field_embedder import MLMTextFieldEmbedder
