--- t2t/modules/token_embedders/__init__.py
@@ -1,3 +1,0 @@
-from t2t.modules.token_embedders.pretrained_transformer_embedder_mlm import (
-    PretrainedTransformerEmbedderMLM,
-)
